pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "notion-mpm"
version = "0.1.0"
description = "MCP server for Notion workspace integration"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.3.1",
    "click>=8.1.0",
    "pydantic>=2.10.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "anyio>=4.0.0",
]

[project.scripts]
notion-mpm = "notion_mpm.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/notion_mpm"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
strict = true
python_version = "3.10"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.9.0",
    "mypy>=1.13.0",
    "twine>=6.0.0",
]
